CoreEventCallback concurrency issue.
CoreEventCallback is a class providing callback mechanisms for the core
object. For example, when core.setPositionUm(...) is called, the
corresponding stage axis is moved, and the registered callback function
will be called subsequently. In the callback function, MMStudioMainFrame
will update some gui controls, which are required to be carried out in the
GUI thread.